Alert text changes and Crash scenario fix.
Crash fix for the scenario:
Crashes when you perform a search for a given genus or species. For
example, search under Pinus. I get out 24 species here in Tucson
(including Lupinus). Then, click on a name to get images. When I then
click on the ‘ < Plant-O-Matic’ button in the upper left to return to
the search results and then I click on the next species to view the app
crashes. It does this repeatidly and always after I search for a given
name